fix: adjust handling of readonly method fee coverage
